RemoveOrReplacePluginSelectorTokenContext was RemoveOrReplacePluginSelectorTextComponentContext
